+++ tvsos-backend/routing-python-tool/.gitignore
@@ -2,10 +2,6 @@
 .venv/
 .idea/
 
-# 忽略 dist 文件夹中的所有 .json 文件
-dist/*.json
-
-# 忽略与 .gitignore 同级的 Json 文件夹内的.json 文件
-Json/*.json
+dist/db_config.json
 
 db_config.json
