+++ tvsos-backend/routing-python-tool/poi_type_mapping.json
@@ -1,0 +1,27 @@
+{
+    "加油站": 1,
+    "加气站": 2,
+    "其它能源站": 3,
+    "工厂": 4,
+    "购物相关场所": 5,
+    "家居建材市场": 6,
+    "公司企业": 7,
+    "公司": 8,
+    "体育休闲服务场所": 9,
+    "汽车维修": 10,
+    "汽车综合维修": 11,
+    "东风特约维修": 12,
+    "货车维修": 13,
+    "汽车养护/装饰": 14,
+    "长安汽车维修": 15,
+    "物流速递": 16,
+    "产业园区": 17,
+    "生活服务场所": 18,
+    "货运火车站": 19,
+    "机场货运处": 20,
+    "服装鞋帽皮具店": 21,
+    "综合市场": 22,
+    "商场": 23,
+    "政府机关": 24,
+    "专卖店": 25
+}
